Refactoring, 7-0 Fixes, Calling Uno Changes
Made a few variable name changes to make hopefully make it easier to understand what some of the variables are used for.
Also should have fixed a bug that was causing some players to get an empty hand after playing a 0 card with '7-0 ruleset' enabled
Addded a 'Draw Uno Penalty' button for players who are not using the scripted UNO calling functionality. This is so there is a way for players to draw the 2 card UNO penalty, without having to enable the 'drawing cards per turn' option