7-0 Fixes and Fake Players
In order to help me work out the problems with the 7-0 rules that I've had for a while, I've programmed most of the logic for computer controlled players. I did this in order to test gameplay on my own. They actually work better than I expected them to with minimal work, even in their incomplete state