Bug Fixing , Stacking Rule Changes, and Menu Tweaks
Fixed a syntax error that was probably still allowing multiple players to press the call uno button at the same time and causing the player with uno to draw multiple uno penalties.

Made some changes to the stacking rules to allow players to mix-and match stacking cards if they so choose

made some tweaks to the menu layout.